feat(settings): refactor settings view to own components
