+++ src/types.ts
@@ -109,3 +109,5 @@ export type Feature = keyof WorkspaceApp["features"];
 export type ExecutionLocation = "observer" | "inline";
 
 export type GetFromStorageFn<T> = (storageKey: string) => Promise<T>;
+
+export type CurrentView = "home" | "settings";
